refactor schema: change data types for time, duration, order_index, and repetitions to smaller types

+++ src/main/java/source/code/dto/request/activity/DailyActivityItemCreateDto.java
@@ -18,7 +18,7 @@ import java.time.LocalDate;
 public class DailyActivityItemCreateDto {
     @NotNull
     @Positive
-    private int time;
+    private Short time;
 
     @Positive
     private BigDecimal weight;
