+++ src/main/java/source/code/dto/request/activity/DailyActivityItemUpdateDto.java
@@ -16,7 +16,7 @@ import java.math.BigDecimal;
 public class DailyActivityItemUpdateDto {
     @NotNull
     @Positive
-    private int time;
+    private Short time;
 
     @Positive
     private BigDecimal weight;
